+++ packages/rspress-plugin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/rspress-plugin
 
+## 0.17.0
+
+### Patch Changes
+
+- @module-federation/enhanced@0.17.0
+- @module-federation/rsbuild-plugin@0.17.0
+- @module-federation/sdk@0.17.0
+- @module-federation/error-codes@0.17.0
+
 ## 0.16.0
 
 ### Patch Changes